Move access requirement mocks into new folder
