+++ packages/synapse-react-client/src/components/AccessRequirementList/ManagedACTAccessRequirementRequestFlow/ManagedACTAccessRequirementItem.test.tsx
@@ -11,7 +11,7 @@ import {
 import {
   mockManagedACTAccessRequirement,
   mockManagedACTAccessRequirementWikiPageKey,
-} from '../../../mocks/mockAccessRequirements'
+} from '../../../mocks/accessRequirement/mockAccessRequirements'
 import { SynapseClient, SynapseContextType } from '../../../index'
 import { MOCK_USER_ID } from '../../../mocks/user/mock_user_profile'
 import userEvent from '@testing-library/user-event'
