+++ packages/synapse-react-client/src/components/AccessRequirementList/RequirementItem/SelfSignAccessRequirementItem.test.tsx
@@ -12,7 +12,7 @@ import {
   mockSelfSignAccessRequirement,
   mockSelfSignAccessRequirementWikiPageKey,
   mockToUAccessRequirement,
-} from '../../../mocks/mockAccessRequirements'
+} from '../../../mocks/accessRequirement/mockAccessRequirements'
 import userEvent from '@testing-library/user-event'
 import { MarkdownSynapseProps } from '../../Markdown'
 import SelfSignAccessRequirementItem, {
