+++ packages/synapse-react-client/src/components/SetManagedAccessRequirementFields/SetManagedAccessRequirementsFields.stories.tsx
@@ -1,7 +1,7 @@
 import { Button, Paper } from '@mui/material'
 import { Meta, StoryObj } from '@storybook/react'
 import React, { useRef, useState } from 'react'
-import { mockManagedACTAccessRequirement } from '../../mocks/mockAccessRequirements'
+import { mockManagedACTAccessRequirement } from '../../mocks/accessRequirement/mockAccessRequirements'
 import {
   SetManagedAccessRequirementFields,
   SetManagedAccessRequirementFieldsHandle,
