+++ packages/synapse-react-client/src/components/WikiMarkdownEditorButton/WikiMarkdownEditorButton.test.tsx
@@ -3,7 +3,7 @@ import { render, screen, waitFor, within } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import React from 'react'
 import { MOCK_ACCESS_TOKEN } from '../../mocks/MockSynapseContext'
-import { mockToUAccessRequirementWithWikiPageKey } from '../../mocks/mockAccessRequirements'
+import { mockToUAccessRequirementWithWikiPageKey } from '../../mocks/accessRequirement/mockAccessRequirements'
 import {
   mockEntityRootWikiPage,
   mockEntityWikiPage,
