+++ packages/synapse-react-client/src/mocks/dataaccess/mockAccessRequest.ts
@@ -1,8 +1,8 @@
-import { ACT_TEAM_ID } from '../utils/SynapseConstants'
+import { ACT_TEAM_ID } from '../../utils/SynapseConstants'
 import { AccessType, SubmissionState } from '@sage-bionetworks/synapse-types'
 import { SubmissionSearchResponse } from '@sage-bionetworks/synapse-types'
-import { mockManagedACTAccessRequirement } from './mockAccessRequirements'
-import { MOCK_USER_ID, MOCK_USER_ID_2 } from './user/mock_user_profile'
+import { mockManagedACTAccessRequirement } from '../accessRequirement/mockAccessRequirements'
+import { MOCK_USER_ID, MOCK_USER_ID_2 } from '../user/mock_user_profile'
 
 export const mockSubmissionSearchResponse: SubmissionSearchResponse = {
   results: [
